\name{AlfaroEtAlUnpublished}
\alias{AlfaroEtAlUnpublished}
\docType{data}
\title{
Alfaro et al. unpublished tree of fish
}
\description{
This has a single timetree of fish (AlfaroEtAlUnpublished_tree). The tree is in phylo (\pkg{ape}) format.
}
\usage{data(AlfaroEtAl2009)}
\details{
Tree is unpublished, contact Mike Alfaro for more information}
\source{

}

\examples{
#not run
#data(AlfaroEtAlUnpublished)


#best tree in ape (phylo) format:
#library(ape)
#fishTree<-AlfaroEtAlUnpublished_tree
#plot the tree
#plot(fishTree,show.node.label=TRUE)

}
\keyword{datasets}
