\name{Heath}
\alias{Heath}
\docType{data}
\title{
Heath et al. unpublished trees
}
\description{
This is a set of 20,000 phylogenetic trees in phylo (ape) format.
}
\usage{data(Heath)}
\details{
Unpublished, please contact Tracy Heath.
}
\source{
Heath et al. unpublished

\examples{
#not run
#data(Heath)

#library(ape)
#treeOne<-Heath$trees[[1]]

#plot the tree
#plot(treeOne,show.node.label=TRUE)

}
\keyword{datasets}
