\name{JaffeEtAl2011}
\alias{JaffeEtAl2011}
\docType{data}
\title{
Jaffe et al. 2011 tree of turtles
}
\description{
This has a single timetree (JaffeEtAl2011_tree) from this paper. The tree is in phylo (\pkg{ape}) format.
}
\usage{data(JaffeEtAl2011)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Jaffe, A. L., G. J. Slater, and M. E. Alfaro. The evolution of island gigantism and body size variation in tortoises and turtles. Biology Letters 7: 558-561.
}

\examples{
#not run
#data(JaffeEtAl2011)
#turtleTree<-JaffeEtAl2011_tree
#plot the tree
#plot(turtleTree,show.node.label=TRUE)
}
\keyword{datasets}
