\name{megaTree}
\alias{megaTree}
\docType{data}
\title{
Giant glomogram of vertebrates. Use at your own risk.
}
\description{
This has a single timetree of ~15,000 vertebrates (megaTree_tree). The tree is in phylo (\pkg{ape}) format.
}
\usage{data(megaTree)}
\details{
Tree is unpublished, contact Jon Eastman for more information}
\source{

}

\examples{
#not run
#data(megaTree)
#mTree<-megaTree_tree

#plot the tree
#plot(mTree,show.node.label=TRUE)
}
\keyword{datasets}
