\name{BergmannEtAl2012}
\alias{BergmannEtAl2012}
\docType{data}
\title{
Bergmann and Irschick 2012 tree of squamates
}
\description{
This has a single timetree (BergmannEtAl2012_tree) from this paper. The tree is in phylo (ape) format.
}
\usage{data(BergmannEtAl2012)}
\details{
Make sure to cite the paper below when you use this dataset.
}
\source{
Bergmann, P. J. and D. J. Irschick. Vertebral evolution and the diversification of squamate reptiles. Evolution 66: 1044-1058.
}

\examples{
#not run
#data(BergmannEtAl2012)
#best.phylo4<-BergmannEtAl2012_tree

#plot the tree
#plot(BergmannEtAl2012_tree,show.node.label=TRUE)
}
\keyword{datasets}
