\name{EastmanEtAlUnpublished}
\alias{EastmanEtAlUnpublished}
\docType{data}
\title{
Eastman et al. unpublished tree of salamanders
}
\description{
This has a single timetree of salamanders (EastmanEtAlUnpublished_tree). The tree is in phylo (\pkg{ape}) format.
}
\usage{data(EastmanEtAlUnpublished)}
\details{
Tree is unpublished, contact Jon Eastman for more information}
\source{

}

\examples{
#not run
#data(EastmanEtAlUnpublished)
#salamanderTree<-EastmanEtAlUnpublished_tree
#plot(salamanderTree,show.node.label=TRUE)
}
\keyword{datasets}
